refactor: rename, cleanup, and remove unused import
